+++ lib/lecture/ClassRoom.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:stocodi_app/lecture/widget/ClassRoomLecture.dart';
 import '../theme/ClassRoomLectureTheme.dart';
-import 'Item/CourseItem.dart';
+import 'Item/CourseListItem.dart';
 import 'package:stocodi_app/lecture/data/PredefinedClassRoomData.dart';
 
 final theme = ClassRoomLectureTheme.getAppTheme();
@@ -12,8 +12,6 @@ class ClassRoom extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     final screenHeight = MediaQuery.of(context).size.height;
-    final screenWidth = MediaQuery.of(context).size.width;
-    final isNarrowScreen = screenWidth < 600; // 예시 너비 (조정 가능)
     return Scaffold(
       appBar: AppBar(
         backgroundColor: theme.backgroundColor,
@@ -32,6 +30,7 @@ class ClassRoom extends StatelessWidget {
           children: [
             SizedBox(
               height: screenHeight * 0.37,
+              width: MediaQuery.of(context).size.width,
               child: ClassRoomLecture(),
             ),
             CourseItem(courseTitle: '실시간 인기강의', courseList: courseList),
