+++ lib/lecture/widget/VideoDetails.dart
@@ -17,10 +17,7 @@ class VideoDetails extends StatelessWidget {
         children: [
           Container(
             margin: const EdgeInsets.only(bottom: 16),
-            child: Text(
-              '초보 주식투자자들이 흔히 하는 실수 박곰희 1부',
-              style: textTheme.displayLarge,
-            ),
+            child: Text('초보 주식투자자들이 흔히 하는 실수 박곰희 1부', style: textTheme.displayLarge,),
           ),
           Row(
             crossAxisAlignment: CrossAxisAlignment.center,
